+++ src/picasapy/app/i18n/picasapy_hu.ts
@@ -2686,6 +2686,10 @@ Az arcfelismerés ettől független: egy mappa lehet figyelt úgy is, hogy az ar
         <translation>Mappák a lemezen</translation>
     </message>
     <message>
+        <source>You can drag and drop pictures here to make a new album.</source>
+        <translation>Ide húzva a képeket új album készül.</translation>
+    </message>
+    <message>
         <source>Scanning for faces... %1% complete</source>
         <translation>Arcok keresése… %1% kész</translation>
     </message>